fix: 최신순,인기순 필터링 오류 해결
