+++ src/pages/dashboard/actor/ActorList.tsx
@@ -1,36 +1,56 @@
 import React, {useState, useCallback} from 'react';
 import {useFocusEffect} from '@react-navigation/native';
-import {FlatList, View, ActivityIndicator} from 'react-native';
+import {
+  FlatList,
+  View,
+  ActivityIndicator,
+  Text,
+  TouchableOpacity,
+} from 'react-native';
 import {GetPostList} from '@/api/post.api';
 import ItemPostNoCategory from '@/components/homeList/ItemPostNoCategory';
+import PostStyles from '../PostStyles';
 
-type ActorListProps = {
-  selectedFilter: string;
-};
-
-const ActorList: React.FC<ActorListProps> = ({selectedFilter}) => {
+const ActorList: React.FC = () => {
   const [postList, setPostList] = useState<any[]>([]);
   const [isLoading, setIsLoading] = useState<boolean>(true); // 초기 로딩 상태
   const [isFetching, setIsFetching] = useState<boolean>(false); // 추가 데이터 로딩 상태
   const [hasMore, setHasMore] = useState<boolean>(true); // 마지막 데이터인지 확인
   const [cursor, setCursor] = useState<number | undefined>(undefined);
+  const [modalVisible, setModalVisible] = useState(false);
+  const [selectedFilter, setSelectedFilter] = useState<'최신순' | '인기순'>(
+    '최신순',
+  );
+
+  const handlePress = (filter: '최신순' | '인기순') => {
+    if (selectedFilter !== filter) {
+      setSelectedFilter(filter);
+    }
+  };
 
   const filterMapping: Record<string, string> = {
     최신순: 'createdat',
     인기순: 'likecount',
   };
+
   const fetchPostList = async (reset = false) => {
     const sortFilter = filterMapping[selectedFilter] || 'createdat';
+
     try {
       if (reset) {
         setIsLoading(true);
+        setCursor(undefined);
+        setPostList([]);
       } else {
         setIsFetching(true);
       }
+
+      const requestCursor = reset ? undefined : cursor;
+
       const response = await GetPostList(
         3,
         sortFilter,
-        cursor,
+        requestCursor,
         undefined,
         'ACTOR',
         undefined,
@@ -39,11 +59,21 @@ const ActorList: React.FC<ActorListProps> = ({selectedFilter}) => {
 
       if (postData.length > 0) {
         // 마지막 데이터의 ID를 cursor로 저장
-        setCursor(postData[postData.length - 1].id);
-        // console.log('cursor: ', cursor);
+        const newCursor = postData[postData.length - 1].id;
+
+        if (typeof newCursor === 'number') {
+          setCursor(newCursor); // cursor 업데이트
+        }
+        // console.log('cursor: ', newCursor);
 
         // 기존 리스트에 추가하거나 새로고침으로 초기화
-        setPostList(reset ? postData : [...postList, ...postData]);
+        setPostList(prevList => {
+          const newList = [...prevList, ...postData];
+          const uniqueList = Array.from(
+            new Map(newList.map(item => [item.id, item])).values(),
+          );
+          return uniqueList;
+        });
       } else {
         setHasMore(false); // 더 이상 데이터가 없음을 표시
       }
@@ -60,7 +90,10 @@ const ActorList: React.FC<ActorListProps> = ({selectedFilter}) => {
 
   useFocusEffect(
     useCallback(() => {
-      fetchPostList(true);
+      setPostList([]); // 기존 데이터 초기화
+      setCursor(undefined); // 필터 변경 시 cursor 초기화
+      setHasMore(true); // 필터 변경 시 새 요청 가능하도록 hasMore 초기화
+      setTimeout(() => fetchPostList(true), 0); // cursor 상태가 반영된 후 실행
     }, [selectedFilter]),
   );
 
@@ -72,6 +105,33 @@ const ActorList: React.FC<ActorListProps> = ({selectedFilter}) => {
 
   return (
     <View>
+      <View style={PostStyles.containerCommentTitle}>
+        <Text style={PostStyles.textCategory}>전체글</Text>
+
+        <View style={PostStyles.containerRow}>
+          <TouchableOpacity onPress={() => handlePress('최신순')}>
+            <Text
+              style={[
+                selectedFilter === '최신순'
+                  ? PostStyles.filter
+                  : PostStyles.tab_filter,
+                {marginRight: 12},
+              ]}>
+              최신순
+            </Text>
+          </TouchableOpacity>
+          <TouchableOpacity onPress={() => handlePress('인기순')}>
+            <Text
+              style={[
+                selectedFilter === '최신순'
+                  ? PostStyles.tab_filter
+                  : PostStyles.filter,
+              ]}>
+              인기순
+            </Text>
+          </TouchableOpacity>
+        </View>
+      </View>
       {isLoading ? (
         <ActivityIndicator size="large" />
       ) : (
