+++ src/pages/dashboard/entire/EntireScreen.tsx
@@ -7,16 +7,6 @@ import CarouselHottest from '@/components/carousel/dashboard/CarouselHottest';
 import EntireList from './EntireList';
 
 const EntireScreen: React.FC = () => {
-  const [selectedFilter, setSelectedFilter] = useState<'최신순' | '인기순'>(
-    '최신순',
-  );
-
-  const handlePress = (filter: '최신순' | '인기순') => {
-    if (selectedFilter !== filter) {
-      setSelectedFilter(filter);
-    }
-  };
-
   return (
     <View style={EntireStyles.container}>
       <View style={EntireStyles.notice}>
@@ -39,38 +29,8 @@ const EntireScreen: React.FC = () => {
         <CarouselHottest />
       </View>
 
-      <View style={EntireStyles.fiter_container}>
-        <Text style={EntireStyles.more_title}>더 찾아보기</Text>
-
-        <TouchableOpacity
-          style={{paddingLeft: 162}}
-          onPress={() => handlePress('최신순')}>
-          <Text
-            style={
-              selectedFilter === '최신순'
-                ? EntireStyles.filter
-                : EntireStyles.tab_filter
-            }>
-            최신순
-          </Text>
-        </TouchableOpacity>
-
-        <TouchableOpacity
-          style={{paddingLeft: 11}}
-          onPress={() => handlePress('인기순')}>
-          <Text
-            style={
-              selectedFilter === '인기순'
-                ? EntireStyles.filter
-                : EntireStyles.tab_filter
-            }>
-            인기순
-          </Text>
-        </TouchableOpacity>
-      </View>
-
       <View>
-        <EntireList selectedFilter={selectedFilter} />
+        <EntireList />
       </View>
     </View>
   );
