+++ src/pages/dashboard/review/ReviewList.tsx
@@ -1,20 +1,43 @@
 import React, {useState, useCallback} from 'react';
 import {useFocusEffect} from '@react-navigation/native';
-import {FlatList, View, ActivityIndicator} from 'react-native';
+import {
+  FlatList,
+  View,
+  ActivityIndicator,
+  Text,
+  TouchableOpacity,
+} from 'react-native';
 import {GetPostList} from '../../../api/post.api';
 import ItemPost from '@/components/homeList/ItemPost';
+import {SvgXml} from 'react-native-svg';
+import {DashboardIcon} from '@/assets/icons/dashboard/DashboardIcon';
+import ModalCategory from '@/components/categoryModal/ModalCategory';
+import PostStyles from '../PostStyles';
 
-type ReviewListProps = {
-  selectedFilter: string;
-  category: string;
-};
-
-const ReviewList: React.FC<ReviewListProps> = ({selectedFilter, category}) => {
+const ReviewList: React.FC = () => {
   const [postList, setPostList] = useState<any[]>([]);
   const [isLoading, setIsLoading] = useState<boolean>(true); // 초기 로딩 상태
   const [isFetching, setIsFetching] = useState<boolean>(false); // 추가 데이터 로딩 상태
   const [hasMore, setHasMore] = useState<boolean>(true); // 마지막 데이터인지 확인
   const [cursor, setCursor] = useState<number | undefined>(undefined);
+  const [modalVisible, setModalVisible] = useState(false);
+  const [category, setCategory] = useState('카테고리 선택');
+  const categoryList = ['전체보기', '시야 후기', '굿즈 후기', '공연 감상'];
+  const [selectedFilter, setSelectedFilter] = useState<'최신순' | '인기순'>(
+    '최신순',
+  );
+
+  const handlePress = (filter: '최신순' | '인기순') => {
+    if (selectedFilter !== filter) {
+      setSelectedFilter(filter);
+    }
+  };
+  const [modalTitle, setModalTitle] = useState('');
+
+  const pressCategory = () => {
+    setModalVisible(true);
+    setModalTitle('카테고리');
+  };
 
   const filterMapping: Record<string, string> = {
     최신순: 'createdat',
@@ -35,13 +58,18 @@ const ReviewList: React.FC<ReviewListProps> = ({selectedFilter, category}) => {
     try {
       if (reset) {
         setIsLoading(true);
+        setCursor(undefined);
+        setPostList([]);
       } else {
         setIsFetching(true);
       }
+
+      const requestCursor = reset ? undefined : cursor;
+
       const response = await GetPostList(
         3,
         sortFilter,
-        cursor,
+        requestCursor,
         sortCategory,
         'REVIEW',
         undefined,
@@ -50,11 +78,22 @@ const ReviewList: React.FC<ReviewListProps> = ({selectedFilter, category}) => {
 
       if (postData.length > 0) {
         // 마지막 데이터의 ID를 cursor로 저장
-        setCursor(postData[postData.length - 1].id);
-        // console.log('cursor: ', cursor);
+        const newCursor = postData[postData.length - 1].id;
+
+        if (typeof newCursor === 'number') {
+          setCursor(newCursor); // cursor 업데이트
+        }
+        // console.log('cursor: ', newCursor);
+        // console.log(postData);
 
         // 기존 리스트에 추가하거나 새로고침으로 초기화
-        setPostList(reset ? postData : [...postList, ...postData]);
+        setPostList(prevList => {
+          const newList = [...prevList, ...postData];
+          const uniqueList = Array.from(
+            new Map(newList.map(item => [item.id, item])).values(),
+          );
+          return uniqueList;
+        });
       } else {
         setHasMore(false); // 더 이상 데이터가 없음을 표시
       }
@@ -71,10 +110,10 @@ const ReviewList: React.FC<ReviewListProps> = ({selectedFilter, category}) => {
 
   useFocusEffect(
     useCallback(() => {
-      setCursor(undefined); // 카테고리가 변경될 때 초기화
       setPostList([]); // 기존 데이터 초기화
-      setHasMore(true); // 새 요청 가능하도록 초기화
-      fetchPostList(true);
+      setCursor(undefined); // 필터 변경 시 cursor 초기화
+      setHasMore(true); // 필터 변경 시 새 요청 가능하도록 hasMore 초기화
+      setTimeout(() => fetchPostList(true), 0); // cursor 상태가 반영된 후 실행
     }, [selectedFilter, category]),
   );
 
@@ -86,6 +125,46 @@ const ReviewList: React.FC<ReviewListProps> = ({selectedFilter, category}) => {
 
   return (
     <View>
+      <View style={PostStyles.containerCommentTitle}>
+        <TouchableOpacity
+          style={PostStyles.containerRow}
+          onPress={() => pressCategory()}>
+          <Text style={PostStyles.textCategory}>{category}</Text>
+          <SvgXml xml={DashboardIcon.arrowDown} />
+        </TouchableOpacity>
+        <ModalCategory
+          modalVisible={modalVisible}
+          setModalVisible={setModalVisible}
+          categoryList={categoryList}
+          modalTitle={modalTitle}
+          onSelect={(item: string) => {
+            setCategory(item);
+          }}
+        />
+        <View style={PostStyles.containerRow}>
+          <TouchableOpacity onPress={() => handlePress('최신순')}>
+            <Text
+              style={[
+                selectedFilter === '최신순'
+                  ? PostStyles.filter
+                  : PostStyles.tab_filter,
+                {marginRight: 12},
+              ]}>
+              최신순
+            </Text>
+          </TouchableOpacity>
+          <TouchableOpacity onPress={() => handlePress('인기순')}>
+            <Text
+              style={[
+                selectedFilter === '최신순'
+                  ? PostStyles.tab_filter
+                  : PostStyles.filter,
+              ]}>
+              인기순
+            </Text>
+          </TouchableOpacity>
+        </View>
+      </View>
       {isLoading ? (
         <ActivityIndicator size="large" />
       ) : (
